@@ -1,14 +1,13 @@
 package com.fraserbrooks.progresstracker.calendar;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.view.ViewPager;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.viewpager.widget.ViewPager;
 import android.util.AttributeSet;
 import android.view.View;
 
 public class CalendarViewPager extends ViewPager{
-
 
     public CalendarViewPager(@NonNull Context context) {
         super(context);
