@@ -2,9 +2,8 @@ package com.fraserbrooks.progresstracker.util;
 
 import java.util.concurrent.Executor;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
-import java.util.concurrent.Executor;
 import java.util.concurrent.Executors;
 
 /**
