@@ -15,7 +15,7 @@
  */
 package com.fraserbrooks.progresstracker.util;
 
-import android.support.test.espresso.IdlingResource;
+import androidx.test.espresso.IdlingResource;
 
 /**
  * Contains a static reference to {@link IdlingResource}, only available in the 'mock' build type.
