@@ -16,7 +16,7 @@
 
 package com.fraserbrooks.progresstracker.util;
 
-import android.support.test.espresso.IdlingResource;
+import androidx.test.espresso.IdlingResource;
 
 import java.util.concurrent.atomic.AtomicInteger;
 
